backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adms-backend"
version = "0.1.0"
description = "Archival Document Management System - Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "aiomysql>=0.2.0",
    "alembic>=1.13.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "python-magic>=0.4.27",
    "pytesseract>=0.3.10",
    "pikepdf>=9.0.0",
    "Pillow>=10.3.0",
    "pdf2image>=1.17.0",
    "lxml>=5.2.0",
    "httpx>=0.27.0",
    "bleach>=6.1.0",
    "citeproc-py>=0.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "black>=24.4.0",
    "mypy>=1.10.0",
]
llm = [
    "openai>=1.30.0",
    "anthropic>=0.28.0",
]
ner = [
    "spacy>=3.7.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM"]

[tool.black]
target-version = ["py312"]
line-length = 100

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
